Rename main app script
